pyproject.toml
[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | _build
    | build
    | dist
  )/
  | requirements.txt
                     # the root of the project
)
'''

[tool.poetry]
name = "elastic-irc"
version = "1.0.0"
description = ""
authors = ["EdLeafe <ed@leafe.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.26.0"
python-dateutil = "^2.8.2"
pymysql = "^1.1.0"
rich = "^13.7.0"
click = "^8.1.7"
elasticsearch7 = "^7.17.9"
requests = "^2.27.1"

[tool.poetry.group.dev.dependencies]
pytest-pudb = "^0.7.0"
ipython = "^8.21.0"
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
